refactor: use .default() to set default value of zod schema (#8039)
### What changed?
- Use `.default()` to set the default value of zod schema.
- Refactor `_lastProps` initialize logic in `edit-session`.
- Move zod schemas to separate `zod-schema` file.
- Move consts to `model/consts` folder.
