+++ packages/affine/model/src/consts/index.ts
@@ -1,3 +1,4 @@
+export * from './connector.js';
 export * from './line.js';
 export * from './note.js';
 export * from './shape.js';
