+++ packages/affine/model/src/elements/connector/local-connector.ts
@@ -3,16 +3,15 @@ import type { SerializedXYWH } from '@blocksuite/global/utils';
 
 import { GfxLocalElementModel } from '@blocksuite/block-std/gfx';
 
+import type { PointStyle } from '../../consts/index.js';
+import type { Connection } from './connector.js';
+
 import {
   type Color,
+  ConnectorMode,
   DEFAULT_ROUGHNESS,
   StrokeStyle,
 } from '../../consts/index.js';
-import {
-  type Connection,
-  ConnectorMode,
-  type PointStyle,
-} from './connector.js';
 
 export class LocalConnectorElementModel extends GfxLocalElementModel {
   private _path: PointLocation[] = [];
