+++ packages/affine/shared/src/utils/index.ts
@@ -11,3 +11,4 @@ export * from './reordering.js';
 export * from './string.js';
 export * from './title.js';
 export * from './url.js';
+export * from './zod-schema.js';
