+++ packages/blocks/src/index.ts
@@ -80,7 +80,6 @@ export {
   MindmapService,
   MindmapStyle,
   MindmapSurfaceBlock,
-  type PointStyle,
   ShapeElementModel,
   SurfaceBlockModel,
   TextElementModel,
