+++ packages/blocks/src/root-block/edgeless/components/panel/line-styles-panel.ts
@@ -3,12 +3,11 @@ import {
   DashLineIcon,
   StraightLineIcon,
 } from '@blocksuite/affine-components/icons';
-import { StrokeStyle } from '@blocksuite/affine-model';
+import { type LineWidth, StrokeStyle } from '@blocksuite/affine-model';
 import { html } from 'lit';
 import { classMap } from 'lit/directives/class-map.js';
 import { repeat } from 'lit/directives/repeat.js';
 
-import type { LineWidth } from '../../../../_common/types.js';
 import type { LineWidthEvent } from './line-width-panel.js';
 
 export type LineStyleEvent =
