+++ packages/blocks/src/root-block/edgeless/components/toolbar/brush/brush-tool-button.ts
@@ -2,7 +2,7 @@ import {
   EdgelessPenDarkIcon,
   EdgelessPenLightIcon,
 } from '@blocksuite/affine-components/icons';
-import { DEFAULT_BRUSH_COLOR } from '@blocksuite/affine-model';
+import { DEFAULT_BRUSH_COLOR, LineWidth } from '@blocksuite/affine-model';
 import { ThemeObserver } from '@blocksuite/affine-shared/theme';
 import { LitElement, css, html } from 'lit';
 import { customElement, state } from 'lit/decorators.js';
@@ -10,7 +10,6 @@ import { styleMap } from 'lit/directives/style-map.js';
 
 import type { LastProps } from '../../../../../surface-block/managers/edit-session.js';
 
-import { LineWidth } from '../../../../../_common/utils/index.js';
 import '../../buttons/toolbar-button.js';
 import { getTooltipWithShortcut } from '../../utils.js';
 import {
