+++ packages/blocks/src/root-block/edgeless/components/toolbar/connector/connector-tool-button.ts
@@ -4,6 +4,7 @@ import {
 } from '@blocksuite/affine-components/icons';
 import {
   DEFAULT_CONNECTOR_COLOR,
+  LineWidth,
   getConnectorModeName,
 } from '@blocksuite/affine-model';
 import { LitElement, css, html } from 'lit';
@@ -12,7 +13,6 @@ import { styleMap } from 'lit/directives/style-map.js';
 
 import type { LastProps } from '../../../../../surface-block/managers/edit-session.js';
 
-import { LineWidth } from '../../../../../_common/utils/index.js';
 import { ConnectorMode } from '../../../../../surface-block/index.js';
 import '../../buttons/toolbar-button.js';
 import { getTooltipWithShortcut } from '../../utils.js';
